refactor: split headers
