+++ papers/TheWitnessEditor/impl/WitnessCLHandler.cpp
@@ -1,4 +1,4 @@
-#include "../Driver.h"
+#include "Driver.h"
 
 int WitnessCLHandler(char * /*argument*/[], int /*maxNumArgs*/)
 {
